Questionable changes added on Saturday at Week 0 event
This was able to run the elevator up and down using the SetControl
with the DutyCycleOut parameter in an open loop; however, we were
not able to get the motors to work together when trying to use
SetControl with a specific position. The two motors were fighting
each other instead of having motor 2 follow motor 1.